@@ -13,6 +13,8 @@ import {
   ɵPromiseProxy
 } from '@angular/fire';
 import { analytics } from 'firebase/app';
+import { registerAnalytics } from '@firebase/analytics';
+import firebase from '@firebase/app';
 
 export interface Config {
   [key: string]: any;
@@ -64,8 +66,8 @@ export class AngularFireAnalytics {
 
     if (!analyticsInitialized) {
       if (isPlatformBrowser(platformId)) {
-        gtag = window[GTAG_FUNCTION_NAME] || ((...args: any[]) => {
-          window[DATA_LAYER_NAME].push(args);
+        gtag = (window[GTAG_FUNCTION_NAME] as any) || ((...args: any[]) => {
+          (window[DATA_LAYER_NAME] as any).push(args);
         });
         window[DATA_LAYER_NAME] = window[DATA_LAYER_NAME] || [];
         analyticsInitialized = zone.runOutsideAngular(() =>
@@ -91,7 +93,12 @@ export class AngularFireAnalytics {
         observeOn(new ɵAngularFireSchedulers(zone).outsideAngular),
         switchMap(() => isPlatformBrowser(platformId) ? import('firebase/analytics') : EMPTY),
         map(() => ɵfirebaseAppFactory(options, zone, nameOrConfig)),
-        map(app => app.analytics()),
+        map(app => {
+          if (registerAnalytics) {
+            registerAnalytics(firebase as any);
+          }
+          return app.analytics();
+        }),
         tap(analytics => {
           if (analyticsCollectionEnabled === false) {
             analytics.setAnalyticsCollectionEnabled(false);
